+++ internal/sdk/plugins.go
@@ -16,11 +16,9 @@ import (
 )
 
 // Plugins - A plugin entity represents a plugin configuration that will be executed during the HTTP request/response lifecycle. Plugins let you add functionality to services that run behind a Kong Gateway instance, like authentication or rate limiting.
-// You can find more information about available plugins and which values each plugin accepts at the [Plugin Hub](https://docs.konghq.com/hub/).
+// You can find more information about available plugins and which values each plugin accepts at the [Plugin Hub](https://developer.konghq.com/plugins/).
 // <br><br>
 // When adding a plugin configuration to a service, the plugin will run on every request made by a client to that service. If a plugin needs to be tuned to different values for some specific consumers, you can do so by creating a separate plugin instance that specifies both the service and the consumer, through the service and consumer fields.
-// <br><br>
-// Plugins can be both [tagged and filtered by tags](https://docs.konghq.com/gateway/latest/admin-api/#tags).
 type Plugins struct {
 	rootSDK          *Konnect
 	sdkConfiguration config.SDKConfiguration
@@ -58842,12 +58840,12 @@ func (s *Plugins) FetchPluginSchema(ctx context.Context, request operations.Fetc
 				return nil, err
 			}
 
-			var out operations.FetchPluginSchemaResponseBody
+			var out shared.GetPluginSchemaResponse
 			if err := utils.UnmarshalJsonFromResponseBody(bytes.NewBuffer(rawBody), &out, ""); err != nil {
 				return nil, err
 			}
 
-			res.Object = &out
+			res.GetPluginSchemaResponse = &out
 		default:
 			rawBody, err := utils.ConsumeRawBody(httpRes)
 			if err != nil {